refactor: use Decimal for monetary fields
- Replace floating-point money fields with rust_decimal::Decimal
  across record, fin_month_record, fin_month_item_record entities
  and their backup mirrors
- Add common/money helper module for shared decimal handling
- Add migration m20260620_000001_money_decimal_text to convert
  existing money columns to text-backed decimal storage
- Enable sea-orm with-rust_decimal feature and add rust_decimal
  dependency with serde-with-str
- Update handlers, models, and export_task service to operate on
  Decimal values for precise financial calculations

+++ mrecord-rust/src/entity/fin_month_item_record.rs
@@ -6,6 +6,7 @@
 //! 数据库表: `FIN_MONTH_ITEM_RECORD`
 
 use chrono::NaiveDateTime;
+use rust_decimal::Decimal;
 use sea_orm::entity::prelude::*;
 use serde::Serialize;
 
@@ -29,7 +30,7 @@ pub struct Model {
     pub template_item_id: String,
     /// 当月该记账项实际金额
     #[sea_orm(column_name = "MR_ITEM_VALUE")]
-    pub item_value: f64,
+    pub item_value: Decimal,
     /// 创建人
     #[sea_orm(column_name = "MR_CREATE_BY")]
     pub create_by: Option<String>,
